+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/block/MagnumTorchBlock.kt
@@ -11,7 +11,11 @@ import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.phys.shapes.CollisionContext
 import net.minecraft.world.phys.shapes.VoxelShape
 
-class MagnumTorchBlock : Block(Properties.ofFullCopy(Blocks.OAK_PLANKS)), EntityBlock {
+class MagnumTorchBlock : Block(
+	Properties
+		.ofFullCopy(Blocks.OAK_PLANKS)
+		.lightLevel { 14 }
+), EntityBlock {
 
 	override fun newBlockEntity(pos: BlockPos, state: BlockState): BlockEntity {
 		return MagnumTorchBlockEntity(pos, state)